backend/requirements.txt
fastapi
uvicorn
fastf1
pandas
numpy
pydantic
langchain
langgraph
openai
python-dotenv
duckduckgo-search
sentence-transformers
pinecone-client
networkx
matplotlib
